Working version with interface changes
according to new design

+++ GarageSaleApp/MenuViewController.m
@@ -68,12 +68,12 @@
 #pragma mark Table View Delegate Methods
 
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    return self.menuItems.count;
+    return  self.menuItems.count;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 80.0;
 }
@@ -170,7 +170,7 @@
 {
     NSLog(@"You are logged in!");
     
-    //[self requestPermissions];
+    // [self requestPermissions];
 
 }
